requirements.txt
python-telegram-bot==13.15
python-binance
python-dotenv
matplotlib
pandas